requirements.txt
discord.py
python-dotenv
imageio
pillow